% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_dfItems}
\alias{get_dfItems}
\alias{get_dfItemsrev}
\title{Get certain objects needed internally by other PROscorerTools functions}
\usage{
get_dfItems(df, items)

get_dfItemsrev(df, dfItems, revitems, minmax)
}
\arguments{
\item{df}{A data frame given as the argument to \code{\link{scoreScale}}}

\item{items}{Item index, given as argument to \code{\link{scoreScale}}}

\item{dfItems}{A data frame with only items, created and used by
\code{\link{scoreScale}} as an interim step in scoring a scale}

\item{revitems}{Items to reverse, given as argument to
\code{\link{scoreScale}}}

\item{minmax}{Minimum and maximum possible item values, given as argument to
\code{\link{scoreScale}}}
}
\value{
These functions return a data frame containing only the items to be
  scored. In the case of \code{get_dfItemsrev}, the specified items will be
  reverse scored in the returned data frame.
}
\description{
These functions are used internally by other PROscorerTools
  functions, particularly \code{\link{scoreScale}}, to get interim objects
  needed for \code{\link{scoreScale}} to do its job.  Only developers wishing
  to contribute to the \code{\link{PROscorerTools}} package should use these
  functions.  Even developers should avoid using them for anything but
  experimentation.  If you use these functions, be warned that they will
  likely change in future versions of the package in ways that may not be
  compatible with your usage.  They will also be made invisible to users in
  future version of the \code{\link{PROscorerTools}} package. The only reason they
  are visible to users in these initial versions of the package is to make
  the internals of the package more transparent to potential developers to
  facilitate the development and improvement of the package.
}
\keyword{internal}

