% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROscorerTools-package.r
\docType{package}
\name{PROscorerTools}
\alias{PROscorerTools}
\alias{PROscorerTools-package}
\title{PROscorerTools}
\description{
Tools to score Patient-Reported Outcome (PRO), Quality
  of Life (QoL), and other psychometric measures and
  questionnaire-based instruments.
}
\details{
Provides a set of reliable and flexible tools to score PRO, QoL, and
  other psychometric measures.  The scoreScale function can be used to score
  single-scale measures.  Additionally, it and the other PROscorerTools
  functions can be used together to write new functions that score more
  complex measures. Scoring procedures represent a major source of error in
  research studies that rely upon PRO and similar measures.  These errors
  typically go unnoticed, hidden, and/or ignored, eroding the scientific
  integrity of the research and hindering progress in the numerous scientific
  fields that conduct studies that use these measures.  One small scoring
  error can severely compromise measurement validity and reliability, as well
  as make research results impossible to reproduce and replicate. The lofty
  goal of the PROscorerTools package is to eliminate this serious deficiency
  in PRO-based research by serving as the gold-standard repository of scoring
  syntax for PRO-like measures commonly used in research.  Eventually, use of
  PROscorerTools to score outcome measures should be a basic requirement for
  a study to be funded and its results published (in my humble opinion).
}

