% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dfItems.R
\name{get_dfItems}
\alias{get_dfItems}
\alias{get_dfItemsrev}
\title{Get a data frame with only items from user input}
\usage{
get_dfItems(df, items)

get_dfItemsrev(df, dfItems, revitems, minmax)
}
\arguments{
\item{df}{A data frame given as the argument to \code{\link{scoreScale}}}

\item{items}{Item index, given as argument to \code{\link{scoreScale}}}

\item{dfItems}{A data frame with only items, created and used by
\code{\link{scoreScale}} as an interim step in scoring a scale}

\item{revitems}{Items to reverse, given as argument to
\code{\link{scoreScale}}}

\item{minmax}{Minimum and maximum possible item values, given as argument to
\code{\link{scoreScale}}}
}
\value{
These functions return a data frame containing only the items to be
scored. In the case of \code{get_dfItemsrev}, the specified items will be
reverse scored in the returned data frame.
}
\description{
Given a data frame and an item index, returns a data frame
containing only the items.  These functions are used internally by other
PROscorerTools functions, particularly \code{\link{scoreScale}}.  Their job
is to return a data frame containing only the items to be scored.  These
functions are also used in the scoring functions in the
\pkg{PROscorer} package to help process the user's input.  These
functions will be of interest mainly to developers wishing to contribute to
the \pkg{PROscorer} package.
}
